requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
Flask-Cors==3.0.10
Flask-Login==0.6.2
Werkzeug==2.2.3
python-dotenv==0.19.0
PyJWT==2.1.0
bcrypt==3.2.0 